Finish epsilon-optimization solvers; add skeleton structure
This commit does two things: it finishes the epsilon-optimization solvers in the EpsilonOptimization helper submodule, and it adds some skeleton files to the root src folder for inclusion in the QuantumStateTransfer module definition.

Next steps are to actually apply the optimization methods to various state transfer objective functions.

(Of course, several documentation steps remain 'TODO', but we shall focus on that in later stages of the project.)