Finalized (tentatively) state transfer code; updated docs (#5)
We finalize the code for state transfer, one of the three phenomena on quantum networks which this package investigates (the other two being uniform mixing and fractional revival, although we may or may not add more in the future). Docstrings are set up but not yet finalized; and on that note, we update the (also unfinished) docs in the EpsilonOptimization submodule as well.