Move files into core subdir; extract fidelity bounds logic
This commit does two things: first, it moves the files defining the core functions for the three (so far) phenomena investigated--namely state transfer between vertices/pairs of vertices, uniform mixing, and fractional revival--into a core subdirectory of src so that they are not mixed up with the more 'fundamental' files like types.jl, utils.jl, etc.

Second, it extracts the logic for bounding the state transfer fidelity function's derivatives (both first and second) from the state transfer file to utils.jl. This is both a separation-of-concerns choice and a desire to reuse the bounds for uniform mixing and fractional revival implementations.